feat: show days of adjacent months (#1)
* feat: fill days of prev and next month in calendar

Instead of only using empty spaces, it would be cool to actually show days of the previous/next month. I implemented a method getPrefill & getPostfill and used them in the njk template. Also both prefill & postfill got their own config value (which defaults to true). They return either empty lists if the setting was set to false or return the days of the prev/next month if the setting was set to true.

* fix: merge prefill and postfill into one option

Now we can either show preview for both, the previous and next month, or none at all.